Preparing to allow Control to safely interrupt Execution threads.
It's a lot more involved than just slamming an interrupt, since we do
not want to risk throwing an interrupt only to have it race and land
just after execution finishes, thus killing the Execution thread.

So we need to basically halt the execution thread, inspect it, and then
decide if it can move on, "Papers, Please"-style.