Testing refactoring for execution interrupt/interdiction
This is pretty experimental, and much more complicated than simply throwing
a KeyboardInterrupt at it. Specifically we really need to make sure that
the thread is the correct one in the correct state. Getting the right one
seems pretty straight forward, but the interrupt/restart mechanics get
wonky once the kernel starts automatically loading the roles...