Reverting a bit to make recovery work better
The way threads are killed and restarted don't quite jive with the context.
The kernel attempts to reuse the same context, which is good, but that
leads to reinitializing and weird bits with (I think) the curve crypto
getting re-set. It also chokes when restarting, since the threads might
still be running, but the engine attempts to add them back in, despite
not being a role yet, leading to weird complaints about it not working.