Dev/release prep (#1)
* Ensure the `dist` folder is always available (for un/packing exports)

* Moving all project resources into a resources folder to better isolate

* Ensure the `build` folder is always available and project is consistent

* Update docs and packers to match resource dir

* Cleanup repo `.gitignore`

* Upgrading Context

* upgrading logger for better designer utility

* SidecarContext added to enable kernels in designers

* missed from Context upgrade...

* now simply unpacks the most recent zip

turns out the workflow is usually "export and unpack" and not a lot of
naming shenanigans. Ideally it'd check the project metadata for a version
number, but... meh

* Fix for hotloader so crypto can be loaded in for encrypted sockets

Hotloading jars is weirdly tricky. Simply adding them to `sys.path` does
not work. At least not reliably. Nor does merely loading in each jar,
one at a time, it seems. I also found that it was necessary to load each
class to make sure they're, I dunno, registered(?) in the class loader.

There was a Java system change in Java 9 that broke the old recipes for
loading jars at runtime in Jython, too. (The system class loaders aren't
of type URLClassLoader anymore.)

By loading all the jars in the same class loader _and_ grinding over by
loading every class, though, it seems to work well. At least it loaded
and ran the curve library from the zmq jar.

* Pass one of adding flavor text to the modules

* Added public key encryption for sockets; fixed hotloading of crypto

* Pass two of adding flavor text to the modules

* Return True on "don't propagate error", not merely None or `pass`

:facepalm:

* Minor refactoring to give subclasses more agency.

This cleans up some of the override bits cluttering the core class def.
Now the JupyterKernelBaseMixin shows exactly what's available for stuff
to implement and cooperate with.

* Preparing to allow Control to safely interrupt Execution threads.

It's a lot more involved than just slamming an interrupt, since we do
not want to risk throwing an interrupt only to have it race and land
just after execution finishes, thus killing the Execution thread.

So we need to basically halt the execution thread, inspect it, and then
decide if it can move on, "Papers, Please"-style.

* Testing refactoring for execution interrupt/interdiction

This is pretty experimental, and much more complicated than simply throwing
a KeyboardInterrupt at it. Specifically we really need to make sure that
the thread is the correct one in the correct state. Getting the right one
seems pretty straight forward, but the interrupt/restart mechanics get
wonky once the kernel starts automatically loading the roles...

* Reverting a bit to make recovery work better

The way threads are killed and restarted don't quite jive with the context.
The kernel attempts to reuse the same context, which is good, but that
leads to reinitializing and weird bits with (I think) the curve crypto
getting re-set. It also chokes when restarting, since the threads might
still be running, but the engine attempts to add them back in, despite
not being a role yet, leading to weird complaints about it not working.

* Restart continuity fixes

Before, I think a lot was owned by the context thread and/or process,
and it wasn't a problem because it didn't matter which thread was the
holding thread. But with more emphasis on using signals to affect threads,
it seems a bit like process is doing context's job... which makes a kind
of sense since that's the Control handler - the one that instigates
restarts and execution context replacements =/

* Fixing listing instances, even across parallel universes via threads

* Corrections from testing project export pack/unpack round trip

* Needs the global props or the WebDev don't authenticate =|

* Tweak for general use